requirements.txt
Pillow
numpy
onnx
onnx_tf
scipy
tensorflow
PySimpleGUI
matplotlib
onnxruntime
pycryptodome